+++ data_privacy_law/tests/test_ui_comprehensivestatepage.py
@@ -10,45 +10,13 @@ This module provides unit tests to verify that the State Comprehensive Laws page
 import unittest
 from streamlit.testing.v1 import AppTest
 
-def find_widgets(widget, target_type):
-    """
-    Goal of this function is to automate the searching of attributes (title, table, images etc.) in streamlit.
-    Required as the structure of streamlit pages is unclear (ex: how to access a subheader?)
+from tests.test_utils import find_widgets
 
-    Implemetation is done using a recursive search.
-    Idea is to find a widget (given in argument 1) which matches a particular thing we are looking for (like title) - 
-    which is mentioned in argument 2 of the function
-
-    If a widget has some children which are lists or dictionaries, then we loop through their contents till we find
-    the target type. 
-
-    Args:
-        widget: The widget or container object to search within.
-        target_type (str): The type of widget to search for (e.g., 'subheader', 'markdown', 'imgs').
-
-    Returns:
-        list: A list of widget objects that match the target type.
-    """
-    found = []
-    if hasattr(widget, 'type') and widget.type == target_type:
-        found.append(widget)
-    if hasattr(widget, 'children'):
-        children = widget.children
-        if isinstance(children, dict):
-            for child in children.values():
-                if isinstance(child, list):
-                    for subchild in child:
-                        found.extend(find_widgets(subchild, target_type))
-                else:
-                    found.extend(find_widgets(child, target_type))
-        elif isinstance(children, list):
-            for child in children:
-                found.extend(find_widgets(child, target_type))
-    return found
 
 class StateComprehensivePageTest(unittest.TestCase):
     """
-    Unit tests for verifying subheader and column content elements on the State Comprehensive Laws page.
+    Unit tests for verifying subheader and column content elements on the
+    State Comprehensive Laws page.
     """
 
     def setUp(self):
@@ -63,7 +31,7 @@ class StateComprehensivePageTest(unittest.TestCase):
         Verify that the title of the page matches the expected title.
         """
         self.assertEqual(
-            self.at.title[0].value, 
+            self.at.title[0].value,
             'Explore Comprehensive State Privacy Laws',
             "The page title does not match the expected title."
         )
