+++ data_privacy_law/tests/test_ui_gdpr.py
@@ -10,47 +10,14 @@ This module provides unit tests to verify that the GDPR page UI:
 import unittest
 from streamlit.testing.v1 import AppTest
 
+from tests.test_utils import find_widgets
 
-def find_widgets(widget, target_type):
-    """
-    Goal of this function is to automate the searching of attributes (title, table, images etc.) in streamlit.
-    Required as the structure of streamlit pages is unclear (ex: how to access a subheader?)
-
-    Implemetation is done using a recursive search.
-    Idea is to find a widget (given in argument 1) which matches a particular thing we are looking for (like title) - 
-    which is mentioned in argument 2 of the function
-
-    If a widget has some children which are lists or dictionaries, then we loop through their contents till we find
-    the target type. 
-
-    Args:
-        widget: The widget or container object to search within.
-        target_type (str): The type of widget to search for (e.g., 'subheader', 'markdown', 'imgs').
-
-    Returns:
-        list: A list of widget objects that match the target type.
-    """
-    found = []
-    if hasattr(widget, 'type') and widget.type == target_type:
-        found.append(widget)
-    if hasattr(widget, 'children'):
-        children = widget.children
-        if isinstance(children, dict):
-            for child in children.values():
-                if isinstance(child, list):
-                    for subchild in child:
-                        found.extend(find_widgets(subchild, target_type))
-                else:
-                    found.extend(find_widgets(child, target_type))
-        elif isinstance(children, list):
-            for child in children:
-                found.extend(find_widgets(child, target_type))
-    return found
 
 
 class StateFederalPageSubheaderTest(unittest.TestCase):
     """
-    Unit tests for verifying subheader and column content elements on the Federal Privacy Laws page.
+    Unit tests for verifying subheader and column content elements on the
+    Federal Privacy Laws page.
     """
 
     def setUp(self):
@@ -94,8 +61,9 @@ class StateFederalPageSubheaderTest(unittest.TestCase):
         for widget in self.at:
             markdown_widgets.extend(find_widgets(widget, 'markdown'))
 
-        expected_snippet = "The General Data Protection Regulation (GDPR) is a comprehensive data protection law " \
-        "enacted by the European Union (EU)"
+        expected_snippet = "The General Data Protection Regulation (GDPR)" \
+"is a comprehensive data protection law " \
+"enacted by the European Union (EU)"
 
         self.assertTrue(
             any(expected_snippet in mw.value for mw in markdown_widgets if mw.value),
@@ -118,8 +86,8 @@ class StateFederalPageSubheaderTest(unittest.TestCase):
             len(image_widgets) > 0,
             "Expected an image in the image column, but no image widget was found."
         )
-    
-    
+
+
     def test_title(self):
         """
         Verify to see that the title of the page matches the expected title.
