+++ data_privacy_law/tests/test_ui_stateprivacy.py
@@ -11,7 +11,10 @@ This module provides unit tests to verify that the State Privacy Law app UI:
 
 import unittest
 import importlib.util
+
 from streamlit.testing.v1 import AppTest
+
+from tests.test_utils import find_widgets
 
 # Use importlib to import the module, since its filename starts with a digit.
 spec = importlib.util.spec_from_file_location(
@@ -21,43 +24,6 @@ state_privacy = importlib.util.module_from_spec(spec)
 spec.loader.exec_module(state_privacy)
 generate_page_summary = state_privacy.generate_page_summary
 map_chunk_to_metadata = state_privacy.map_chunk_to_metadata
-
-
-def find_widgets(widget, target_type):
-    """
-    Goal of this function is to automate the searching of attributes (title, table, images etc.) in streamlit.
-    Required as the structure of streamlit pages is unclear (ex: how to access a subheader?)
-
-    Implemetation is done using a recursive search.
-    Idea is to find a widget (given in argument 1) which matches a particular thing we are looking for (like title) -
-    which is mentioned in argument 2 of the function
-
-    If a widget has some children which are lists or dictionaries, then we loop through their contents till we find
-    the target type.
-
-    Args:
-        widget: The widget or container object to search within.
-        target_type (str): The type of widget to search for (e.g., 'subheader', 'markdown', 'imgs').
-
-    Returns:
-        list: A list of widget objects that match the target type.
-    """
-    found = []
-    if hasattr(widget, "type") and widget.type == target_type:
-        found.append(widget)
-    if hasattr(widget, "children"):
-        children = widget.children
-        if isinstance(children, dict):
-            for child in children.values():
-                if isinstance(child, list):
-                    for subchild in child:
-                        found.extend(find_widgets(subchild, target_type))
-                else:
-                    found.extend(find_widgets(child, target_type))
-        elif isinstance(children, list):
-            for child in children:
-                found.extend(find_widgets(child, target_type))
-    return found
 
 
 class StatePrivacyLawAppTest(unittest.TestCase):
@@ -116,60 +82,6 @@ class StatePrivacyLawAppTest(unittest.TestCase):
             ),
             "Question input field with expected label not found.",
         )
-
-    # def test_convert_date_function(self):
-    #     """
-    #     Verify that the convert_date function converts dates correctly.
-    #     """
-
-    #     # Use importlib to import the module, since its filename starts with a digit.
-    #     spec = importlib.util.spec_from_file_location(
-    #         "state_privacy", "app/pages/1_State_Privacy.py"
-    #     )
-    #     state_privacy = importlib.util.module_from_spec(spec)
-    #     spec.loader.exec_module(state_privacy)
-    #     convert_date = state_privacy.convert_date
-
-    #     self.assertEqual(
-    #         convert_date("01012020"),
-    #         "01/01/2020",
-    #         "Date conversion did not work as expected.",
-    #     )
-    #     self.assertEqual(
-    #         convert_date("invalid"),
-    #         "invalid",
-    #         "Non 8-digit string should remain unchanged.",
-    #     )
-
-    def test_generate_page_summary(self):
-        """
-        Test the generate_page_summary function.
-        """
-        # Test valid inputs
-        chunk_ids_with_metadata = [
-            ("./pdfs/Texas/test.pdf", "Texas Privacy Act", "1"),
-            ("./pdfs/Texas/test.pdf", "Texas Privacy Act", "2"),
-        ]
-        user_question = "What are state privacy laws related to social media?"
-        # Test with invalid inputs
-        with self.assertRaises(TypeError):
-            generate_page_summary(None, user_question)
-
-        with self.assertRaises(TypeError):
-            generate_page_summary(chunk_ids_with_metadata, None)
-
-    def test_map_chunk_to_metadata(self):
-        """
-        Test the map_chunk_to_metadata function.
-        """
-        chunk_ids_with_metadata = [
-            ("./pdfs/Texas/test.pdf", "Texas Privacy Act", "1"),
-            ("./pdfs/Texas/test.pdf", "Texas Privacy Act", "2"),
-        ]
-        self.assertTrue(isinstance(chunk_ids_with_metadata, list))
-        with self.assertRaises(TypeError):
-            map_chunk_to_metadata(None)
-
 
 if __name__ == "__main__":
     unittest.main()
